// types/supabase.types.ts
export type Database = {
// ...
    public: {
        Tables: {
// ...
            ingredient_transactions: {
                Row: {
                    created_at: string | null
                    id: string
                    ingredient_id: string | null
                    notes: string | null
                    payment_status: string
                    quantity: number
                    total_cost: number
                    transaction_type: string
                    unit_cost: number
                    updated_at: string | null
                }
                Insert: {
                    created_at?: string | null
                    id?: string
                    ingredient_id?: string | null
                    notes?: string | null
                    payment_status?: string
                    quantity: number
                    total_cost: number
                    transaction_type: string
                    unit_cost: number
                    updated_at?: string | null
                }
                Update: {
                    created_at?: string | null
                    id?: string
                    ingredient_id?: string | null
                    notes?: string | null
                    payment_status?: string
                    quantity?: number
                    total_cost?: number
                    transaction_type?: string
                    unit_cost?: number
                    updated_at?: string | null
                }
                Relationships: [
                    {
                        foreignKeyName: "ingredient_transactions_ingredient_id_fkey"
                        columns: ["ingredient_id"]
                        isOneToOne: false
                        referencedRelation: "ingredients"
                        referencedColumns: ["id"]
                    },
                ]
            }
            ingredients: {
                Row: {
                    category_id: string | null
                    cost: number | null
                    created_at: string | null
                    description: string | null
                    id: string
                    min_stock: number
                    name: string
                    stock: number
                    unit: string
                    updated_at: string | null
                }
                Insert: {
                    category_id?: string | null
                    cost?: number | null
                    created_at?: string | null
                    description?: string | null
                    id?: string
                    min_stock?: number
                    name: string
                    stock?: number
                    unit: string
                    updated_at?: string | null
                }
                Update: {
                    category_id?: string | null
                    cost?: number | null
                    created_at?: string | null
                    description?: string | null
                    id?: string
                    min_stock?: number
                    name?: string
                    stock?: number
                    unit?: string
                    updated_at?: string | null
                }
                Relationships: [
                    {
                        foreignKeyName: "ingredients_category_fkey"
                        columns: ["category_id"]
                        isOneToOne: false
                        referencedRelation: "ingredient_categories"
                        referencedColumns: ["id"]
                    },
                ]
            }
// ...
        }
        Views: {
            [_ in never]: never
        }
        Functions: {
            [_ in never]: never
        }
        Enums: {
            [_ in never]: never
        }
        CompositeTypes: {
            [_ in never]: never
        }
    }
}